scripts/requirements.txt

fastapi 
uvicorn
python-multipart
tensorflow
numpy
pillow
